Move MockedMetricsTest to Diagnostics project
