+++ test/Diagnostics/Metrics/MockedMetricsTest.cs
@@ -8,7 +8,7 @@ using Inspector;
 using Microsoft.ServiceFabric.Diagnostics.Metrics.Implementation;
 using Moq;
 
-namespace Microsoft.ServiceFabric.TestFramework
+namespace Microsoft.ServiceFabric.Diagnostics.Metrics
 {
     /// <summary>
     /// Classes that have a direct or indirect dependency on Metrics API (<see cref="IMeter"/>, <see cref="IMeterProvider"/>) end up calling native metrics Interop which is unavailable during unit testing
@@ -21,10 +21,7 @@ namespace Microsoft.ServiceFabric.TestFramework
             typeof(MeterProvider<long>).Field<Func<IFabricMeterProvider>>().Set(() => new Mock<IFabricMeterProvider>() { DefaultValue = DefaultValue.Mock }.Object);
             typeof(MeterProvider<TimeSpan>).Field<Func<IFabricMeterProvider>>().Set(() => new Mock<IFabricMeterProvider>() { DefaultValue = DefaultValue.Mock }.Object);
         }
-        /// <summary>
-        /// Implementation of <see cref="MeterProvider{TValueType}"/> for long integer metrics.
-        /// Creates meters for recording long integer telemetry values with various dimension configurations.
-        /// </summary>
+
         public virtual void Dispose()
         {
             typeof(MeterProvider<long>).Field<Func<IFabricMeterProvider>>().Set(NativeTelemetry.FabricCreateMeterProvider);
